@@ -1,11 +1,10 @@
 name: circles_background
 description: A simple backround design package to make your app look beautiful.
-version: 0.0.2
+version: 0.0.3
 homepage: https://github.com/dtkdt100/circles_background
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: ^3.9.2
 
 dependencies:
   flutter:
@@ -14,7 +13,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^1.0.0
+  flutter_lints: ^6.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
