@@ -6,15 +6,13 @@ void main() {
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Circles Background',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
+      theme: ThemeData(primarySwatch: Colors.blue),
       debugShowCheckedModeBanner: false,
       home: CirclesBackgroundPage(),
       //home: const ThreeCirclesBackgroundPage(GradientColor.blue),
@@ -25,34 +23,37 @@ class MyApp extends StatelessWidget {
 }
 
 class CirclesBackgroundPage extends StatelessWidget {
-  CirclesBackgroundPage({Key? key}) : super(key: key);
+  CirclesBackgroundPage({super.key});
 
   final List<CircleInfo> circles = [
     CircleInfo(
-        size: const Size(300, 500),
-        color: Colors.green,
-        borderRadius: const BorderRadius.only(bottomLeft: Radius.circular(200)),
-        alignment: Alignment.topRight
+      size: const Size(300, 500),
+      color: Colors.green,
+      borderRadius: const BorderRadius.only(bottomLeft: Radius.circular(200)),
+      alignment: Alignment.topRight,
     ),
     CircleInfo(
-        size: const Size(300, 900),
-        gradient: LinearGradient(
-            begin: Alignment.topCenter,
-            end: Alignment.bottomCenter,
-            colors: [Colors.green[800]!, Colors.green]
-        ),
-        borderRadius: const BorderRadius.only(),
-        alignment: Alignment.topLeft
+      size: const Size(300, 1000),
+      gradient: LinearGradient(
+        begin: Alignment.topCenter,
+        end: Alignment.bottomCenter,
+        colors: [Colors.green[800]!, Colors.green],
+      ),
+      borderRadius: const BorderRadius.only(),
+      alignment: Alignment.topLeft,
     ),
     CircleInfo(
-        size: const Size(200, 500),
-        gradient: LinearGradient(
-            begin: Alignment.topCenter,
-            end: Alignment.bottomCenter,
-            colors: [Colors.green[800]!, Colors.green]
-        ),
-        borderRadius: const BorderRadius.only(bottomLeft: Radius.circular(50), topLeft: Radius.circular(150)),
-        alignment: Alignment.bottomRight
+      size: const Size(200, 500),
+      gradient: LinearGradient(
+        begin: Alignment.topCenter,
+        end: Alignment.bottomCenter,
+        colors: [Colors.green[800]!, Colors.green],
+      ),
+      borderRadius: const BorderRadius.only(
+        bottomLeft: Radius.circular(50),
+        topLeft: Radius.circular(150),
+      ),
+      alignment: Alignment.bottomRight,
     ),
   ];
 
@@ -63,20 +64,21 @@ class CirclesBackgroundPage extends StatelessWidget {
         circles: circles,
         child: const Padding(
           padding: EdgeInsets.symmetric(vertical: 70, horizontal: 25),
-          child: Text('Hello word!', style: TextStyle(
-            color: Colors.white,
-            fontSize: 40
-          ),),
+          child: Text(
+            'Hello word!',
+            style: TextStyle(color: Colors.white, fontSize: 40),
+          ),
         ),
       ),
-
     );
   }
 }
 
 class ThreeCirclesBackgroundPage extends StatelessWidget {
   final GradientColor gradientColor;
-  const ThreeCirclesBackgroundPage(this.gradientColor, {Key? key}) : super(key: key);
+
+  const ThreeCirclesBackgroundPage(this.gradientColor, {Key? key})
+    : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -87,13 +89,12 @@ class ThreeCirclesBackgroundPage extends StatelessWidget {
         sizeOfScreen: MediaQuery.of(context).size,
         child: const Padding(
           padding: EdgeInsets.symmetric(vertical: 70, horizontal: 25),
-          child: Text('Hello word!', style: TextStyle(
-              color: Colors.white,
-              fontSize: 40
-          ),),
+          child: Text(
+            'Hello word!',
+            style: TextStyle(color: Colors.white, fontSize: 40),
+          ),
         ),
       ),
-
     );
   }
 
@@ -105,7 +106,6 @@ class ThreeCirclesBackgroundPage extends StatelessWidget {
         [Colors.pink[700]!, Colors.pink[900]!],
       ];
     }
+    return null;
   }
 }
-
-
